delete duplication
